changement de dossier

+++ georgePompidou/tarification.php
@@ -4,8 +4,7 @@ namespace georgePompidou;
 
 class tarification
 {
-
-    function __construct(public int $prix, public diametre $diametre)
+    function __construct(public float $prix, public diametre $diametre)
     {
     }
